moved build provider dropdown to build component
