+++ client-react/src/pages/app/deployment-center/code/DeploymentCenterCodeSettings.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { DeploymentCenterFieldProps, DeploymentCenterCodeFormData, WorkflowOption } from '../DeploymentCenter.types';
+import { DeploymentCenterFieldProps, DeploymentCenterCodeFormData } from '../DeploymentCenter.types';
 import DeploymentCenterGitHubDataLoader from '../github-provider/DeploymentCenterGitHubDataLoader';
 import { ScmType, BuildProvider } from '../../../../models/site/config';
 import DeploymentCenterCodeBuild from './DeploymentCenterCodeBuild';
@@ -7,20 +7,17 @@ import { DeploymentCenterContext } from '../DeploymentCenterContext';
 import DeploymentCenterGitHubReadOnly from '../github-provider/DeploymentCenterGitHubReadOnly';
 import DeploymentCenterCodeBuildReadOnly from './DeploymentCenterCodeBuildReadOnly';
 import DeploymentCenterGitHubWorkflowConfig from '../github-provider/DeploymentCenterGitHubWorkflowConfig';
-import DeploymentCenterCodeSourceAndBuild from './DeploymentCenterCodeSourceAndBuild';
 import DeploymentCenterGitHubWorkflowConfigSelector from '../github-provider/DeploymentCenterGitHubWorkflowConfigSelector';
+import DeploymentCenterCodeSource from './DeploymentCenterCodeSource';
 
 const DeploymentCenterCodeSettings: React.FC<DeploymentCenterFieldProps<DeploymentCenterCodeFormData>> = props => {
   const { formProps } = props;
   const deploymentCenterContext = useContext(DeploymentCenterContext);
 
+  const isSourceSelected = formProps && formProps.values.sourceProvider !== ScmType.None;
   const isGitHubSource = formProps && formProps.values.sourceProvider === ScmType.GitHub;
   const isGitHubActionsBuild = formProps && formProps.values.buildProvider === BuildProvider.GitHubAction;
   const isDeploymentSetup = deploymentCenterContext.siteConfig && deploymentCenterContext.siteConfig.properties.scmType !== ScmType.None;
-  const isUsingExistingOrAvailableWorkflowConfig =
-    formProps &&
-    (formProps.values.workflowOption === WorkflowOption.UseExistingWorkflowConfig ||
-      formProps.values.workflowOption === WorkflowOption.UseAvailableWorkflowConfigs);
 
   const disconnectCallback = () => {
     throw Error('not implemented');
@@ -35,14 +32,14 @@ const DeploymentCenterCodeSettings: React.FC<DeploymentCenterFieldProps<Deployme
         </>
       ) : (
         <>
-          <DeploymentCenterCodeSourceAndBuild formProps={formProps} />
+          <DeploymentCenterCodeSource />
           {isGitHubSource && (
             <>
               <DeploymentCenterGitHubDataLoader formProps={formProps} />
               {isGitHubActionsBuild && <DeploymentCenterGitHubWorkflowConfigSelector formProps={formProps} />}
             </>
           )}
-          {isGitHubActionsBuild && !isUsingExistingOrAvailableWorkflowConfig && <DeploymentCenterCodeBuild formProps={formProps} />}
+          {isSourceSelected && <DeploymentCenterCodeBuild formProps={formProps} />}
           <DeploymentCenterGitHubWorkflowConfig formProps={formProps} />
         </>
       )}
